Move common fixture to conftest file
